xmlapi: split out xmlbase.py and xmllibxml2.py
We will be adding new XMLAPI implementations shortly and separate
files helps with code org
